# Rework off-heap float32 bulk scoring to reduce garbage creation