--- a/lucene/core/src/java24/org/apache/lucene/internal/vectorization/Lucene99MemorySegmentFloatVectorScorerSupplier.java
+++ b/lucene/core/src/java24/org/apache/lucene/internal/vectorization/Lucene99MemorySegmentFloatVectorScorerSupplier.java
@@ -35,33 +35,34 @@ public abstract sealed class Lucene99MemorySegmentFloatVectorScorerSupplier
   final int vectorByteSize;
   final int maxOrd;
   final int dims;
-  final MemorySegmentAccessInput input;
+  final MemorySegment seg;
   final FloatVectorValues values; // to support ordToDoc/getAcceptOrds
-  byte[] queryScratch, scratch1, scratch2, scratch3, scratch4;
 
   /**
    * Return an optional whose value, if present, is the scorer supplier. Otherwise, an empty
    * optional is returned.
    */
   static Optional<RandomVectorScorerSupplier> create(
-      VectorSimilarityFunction type, IndexInput input, FloatVectorValues values) {
+      VectorSimilarityFunction type, IndexInput input, FloatVectorValues values)
+      throws IOException {
     input = FilterIndexInput.unwrapOnlyTest(input);
-    if (!(input instanceof MemorySegmentAccessInput msInput)) {
+    MemorySegment seg;
+    if (!(input instanceof MemorySegmentAccessInput msInput
+        && (seg = msInput.segmentSliceOrNull(0L, msInput.length())) != null)) {
       return Optional.empty();
     }
     checkInvariants(values.size(), values.getVectorByteLength(), input);
     return switch (type) {
       case COSINE -> Optional.empty(); // of(new CosineSupplier(msInput, values));
-      case DOT_PRODUCT -> Optional.of(new DotProductSupplier(msInput, values));
+      case DOT_PRODUCT -> Optional.of(new DotProductSupplier(seg, values));
       case EUCLIDEAN -> Optional.empty(); // of(new EuclideanSupplier(msInput, values));
       case MAXIMUM_INNER_PRODUCT ->
           Optional.empty(); // of(new MaxInnerProductSupplier(msInput, values));
     };
   }
 
-  Lucene99MemorySegmentFloatVectorScorerSupplier(
-      MemorySegmentAccessInput input, FloatVectorValues values) {
-    this.input = input;
+  Lucene99MemorySegmentFloatVectorScorerSupplier(MemorySegment seg, FloatVectorValues values) {
+    this.seg = seg;
     this.values = values;
     this.vectorByteSize = values.getVectorByteLength();
     this.maxOrd = values.size();
@@ -80,26 +81,15 @@ public abstract sealed class Lucene99MemorySegmentFloatVectorScorerSupplier
     }
   }
 
-  final MemorySegment getSegment(int ord, byte[] scratch) throws IOException {
-    long byteOffset = (long) ord * vectorByteSize;
-    MemorySegment seg = input.segmentSliceOrNull(byteOffset, vectorByteSize);
-    if (seg == null) {
-      if (scratch == null) {
-        scratch = new byte[vectorByteSize];
-      }
-      input.readBytes(byteOffset, scratch, 0, vectorByteSize);
-      seg = MemorySegment.ofArray(scratch);
-    }
-    return seg;
-  }
-
   static final class DotProductSupplier extends Lucene99MemorySegmentFloatVectorScorerSupplier {
 
-    static final MemorySegmentBulkVectorOps.DotFromQuerySegment DOT_OPS =
-        new MemorySegmentBulkVectorOps.DotFromQuerySegment();
+    static final MemorySegmentBulkVectorOps.DotProduct DOT_OPS =
+        MemorySegmentBulkVectorOps.DOT_INSTANCE;
 
-    DotProductSupplier(MemorySegmentAccessInput input, FloatVectorValues values) {
-      super(input, values);
+    final float[] scratchScores = new float[4];
+
+    DotProductSupplier(MemorySegment seg, FloatVectorValues values) {
+      super(seg, values);
     }
 
     @Override
@@ -110,25 +100,24 @@ public abstract sealed class Lucene99MemorySegmentFloatVectorScorerSupplier
         @Override
         public float score(int node) throws IOException {
           checkOrdinal(node);
-          MemorySegment query = getSegment(queryOrd, queryScratch);
-          MemorySegment ms = getSegment(node, scratch1);
-          var raw = DOT_OPS.dotProduct(query, ms, dims);
+          long queryAddr = (long) queryOrd * vectorByteSize;
+          long addr = (long) node * vectorByteSize;
+          var raw = DOT_OPS.dotProduct(seg, queryAddr, addr, dims);
           return normalizeToUnitInterval(raw);
         }
 
         @Override
         public void bulkScore(int[] nodes, float[] scores, int numNodes) throws IOException {
           // TODO checkOrdinal(node1 ....);
-          float[] scratchScores = new float[4];
           int i = 0;
-          MemorySegment query = getSegment(queryOrd, queryScratch);
+          long queryAddr = (long) queryOrd * vectorByteSize;
           final int limit = numNodes & ~3;
           for (; i < limit; i += 4) {
-            MemorySegment ms1 = getSegment(nodes[i], scratch1);
-            MemorySegment ms2 = getSegment(nodes[i + 1], scratch2);
-            MemorySegment ms3 = getSegment(nodes[i + 2], scratch3);
-            MemorySegment ms4 = getSegment(nodes[i + 3], scratch4);
-            DOT_OPS.dotProductBulk(scratchScores, query, ms1, ms2, ms3, ms4, dims);
+            long addr1 = (long) nodes[i + 0] * vectorByteSize;
+            long addr2 = (long) nodes[i + 1] * vectorByteSize;
+            long addr3 = (long) nodes[i + 2] * vectorByteSize;
+            long addr4 = (long) nodes[i + 3] * vectorByteSize;
+            DOT_OPS.dotProductBulk(seg, scratchScores, queryAddr, addr1, addr2, addr3, addr4, dims);
             scores[i + 0] = normalizeToUnitInterval(scratchScores[0]);
             scores[i + 1] = normalizeToUnitInterval(scratchScores[1]);
             scores[i + 2] = normalizeToUnitInterval(scratchScores[2]);
@@ -137,10 +126,10 @@ public abstract sealed class Lucene99MemorySegmentFloatVectorScorerSupplier
           // Handle remaining 1–3 nodes in bulk (if any)
           int remaining = numNodes - i;
           if (remaining > 0) {
-            MemorySegment ms1 = getSegment(nodes[i], scratch1);
-            MemorySegment ms2 = (remaining > 1) ? getSegment(nodes[i + 1], scratch2) : ms1;
-            MemorySegment ms3 = (remaining > 2) ? getSegment(nodes[i + 2], scratch3) : ms1;
-            DOT_OPS.dotProductBulk(scratchScores, query, ms1, ms2, ms3, ms1, dims);
+            long addr1 = (long) nodes[i] * vectorByteSize;
+            long addr2 = (remaining > 1) ? (long) nodes[i + 1] * vectorByteSize : addr1;
+            long addr3 = (remaining > 2) ? (long) nodes[i + 2] * vectorByteSize : addr1;
+            DOT_OPS.dotProductBulk(seg, scratchScores, queryAddr, addr1, addr2, addr3, addr3, dims);
             scores[i] = normalizeToUnitInterval(scratchScores[0]);
             if (remaining > 1) scores[i + 1] = normalizeToUnitInterval(scratchScores[1]);
             if (remaining > 2) scores[i + 2] = normalizeToUnitInterval(scratchScores[2]);
@@ -157,7 +146,7 @@ public abstract sealed class Lucene99MemorySegmentFloatVectorScorerSupplier
 
     @Override
     public DotProductSupplier copy() throws IOException {
-      return new DotProductSupplier(input.clone(), values);
+      return new DotProductSupplier(seg, values);
     }
   }
 }